# Backend/docs/swagger.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Library-API",
        "version": "1.0.0",
        "description": "### This is a Swagger UI for our Library API  \n\n### Error handling\nErrors are presented using the HTTP status codes 400-599. The response include a more detailed explanation about the error in the body.\n\n```\n{\n  \"error\": \"Cleartext explanation of the error that occurred\"\n}\n```\n\n### Generic API errors\nThese error responses can be returned by any of the described endpoints.\n\n| HTTP status  |  Description |\n|---|---|\n| 400| Bad input parameters or invalid request. More information can be found in the response body.\n| 401| The consumer does not have access to the requested resource or the provided token is not valid.\n| 404| The requested resource can not be found or the user does not have access to the resource.\n| 5XX| A server error has occurred while processing the request."
    },
    "paths": {
...
        "/logout/": {
            "post": {
                "tags": [
                    "users"
                ],
                "parameters": [
                    {
                        "examples": {
                            "key": {
                                "value": "30cf8c1d-3fb9-4149-a2eb-20e3e88b89d6"
                            }
                        },
                        "name": "key",
                        "description": "API key is required to access the API",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query",
                        "required": true
                    },
                    {
                        "examples": {
                            "username": {
                                "value": "probookreader123"
                            }
                        },
                        "name": "username",
                        "description": "A users name",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query",
                        "required": true
                    },
                    {
                        "examples": {
                            "password": {
                                "value": "qwerty123456"
                            }
                        },
                        "name": "password",
                        "description": "A password for the users account",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "text/xml": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Logout": {
                                        "value": "Logged out"
                                    }
                                }
                            }
                        },
                        "description": "Request successful"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Missing API key error": {
                                        "value": {
                                            "error": "Missing API key"
                                        }
                                    }
                                }
                            }
                        },
                        "description": "API key is not provided"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Invalid API key error": {
                                        "value": {
                                            "error": "Invalid API key"
                                        }
                                    }
                                }
                            }
                        },
                        "description": "API key is invalid"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Something went wrong error\t": {
                                        "value": {
                                            "error": "Looks like something went wrong :("
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Something went wrong during the query"
                    }
                },
                "summary": "Log a user out"
            }
        },
        "/login/": {
            "post": {
                "tags": [
                    "users"
                ],
                "parameters": [
                    {
                        "examples": {
                            "key": {
                                "value": "30cf8c1d-3fb9-4149-a2eb-20e3e88b89d6"
                            }
                        },
                        "name": "key",
                        "description": "API key is required to access the API",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query",
                        "required": true
                    },
                    {
                        "examples": {
                            "username": {
                                "value": "probookreader123"
                            }
                        },
                        "name": "username",
                        "description": "A users name",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query",
                        "required": true
                    },
                    {
                        "examples": {
                            "password": {
                                "value": "qwerty123456"
                            }
                        },
                        "name": "password",
                        "description": "A password for the users account",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "text/xml": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Login": {
                                        "value": "Logged in"
                                    }
                                }
                            }
                        },
                        "description": "Request successful"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "One or all params are missing error": {
                                        "value": {
                                            "error": "One or all params are missing. Required params: password, username"
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Missing parameters"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Missing API key error": {
                                        "value": {
                                            "error": "Missing API key"
                                        }
                                    }
                                }
                            }
                        },
                        "description": "API key is not provided"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Invalid API key error": {
                                        "value": {
                                            "error": "Invalid API key"
                                        }
                                    }
                                }
                            }
                        },
                        "description": "API key is invalid"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Something went wrong error\t": {
                                        "value": {
                                            "error": "Looks like something went wrong :("
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Something went wrong during the query"
                    }
                },
                "summary": "Log a user in"
            }
        },
        "/users/": {
            "get": {
                "tags": [
                    "users"
                ],
                "parameters": [
                    {
                        "examples": {
                            "key": {
                                "value": "30cf8c1d-3fb9-4149-a2eb-20e3e88b89d6"
                            }
                        },
                        "name": "key",
                        "description": "API key is required to access the API",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                },
                                "examples": {
                                    "Users": {
                                        "value": [
                                            {
                                                "_id": "632849e813c575a0084104fa",
                                                "username": "qwerty"
                                            },
                                            {
                                                "_id": "63284c6c0e4f9ec5457bd09b",
                                                "username": "aaaaaaa"
                                            }
                                        ]
                                    }
                                }
                            }
                        },
                        "description": "Request successful"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Missing API key error": {
                                        "value": {
                                            "error": "Missing API key"
                                        }
                                    }
                                }
                            }
                        },
                        "description": "API key is not provided"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Invalid API key error": {
                                        "value": {
                                            "error": "Invalid API key"
                                        }
                                    }
                                }
                            }
                        },
                        "description": "API key is invalid"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Something went wrong error": {
                                        "value": {
                                            "error": "Looks like something went wrong :("
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Something went wrong during the query"
                    }
                },
                "summary": "Get a list of all users",
                "description": ""
            },
            "post": {
                "tags": [
                    "users"
                ],
                "parameters": [
                    {
                        "examples": {
                            "admin": {
                                "value": "false"
                            }
                        },
                        "name": "admin",
                        "description": "A bool that allows you to create/edit/delete books/loans/users/reviews",
                        "schema": {
                            "type": "boolean"
                        },
                        "in": "query"
                    },
                    {
                        "examples": {
                            "key": {
                                "value": "30cf8c1d-3fb9-4149-a2eb-20e3e88b89d6\t"
                            }
                        },
                        "name": "key",
                        "description": "API key is required to access the API",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query",
                        "required": true
                    },
                    {
                        "examples": {
                            "username": {
                                "value": "probookreader123"
                            }
                        },
                        "name": "username",
                        "description": "A users name",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query",
                        "required": true
                    },
                    {
                        "examples": {
                            "password": {
                                "value": "qwerty123456"
                            }
                        },
                        "name": "password",
                        "description": "A password for the users account",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "content": {
                            "text/xml": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "User": {
                                        "value": "User created!"
                                    }
                                }
                            }
                        },
                        "description": "Request successful"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "One or all params are missing error": {
                                        "value": {
                                            "error": "One or all params are missing. Required params: username, password."
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Missing parameters"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Missing API key error": {
                                        "value": {
                                            "error": "Missing API key"
                                        }
                                    }
                                }
                            }
                        },
                        "description": "API key is not provided"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Invalid API key error": {
                                        "value": {
                                            "error": "Invalid API key"
                                        }
                                    }
                                }
                            }
                        },
                        "description": "API key is invalid"
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Username already is in use error": {
                                        "value": {
                                            "error": "A user with the given username is already registered"
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Username is already taken"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Something went wrong error": {
                                        "value": {
                                            "error": "Looks like something went wrong :("
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Something went wrong during the query"
                    }
                },
                "summary": "Add a new user"
            }
        },
...
    }
}